# docs/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "CSRFHeader": {
        "description": "Signed double-submit CSRF token required for unsafe browser requests in WorkOS auth mode.",
        "in": "header",
        "name": "x-archagent-csrf",
        "type": "apiKey"
      },
      "CloudTasksOIDC": {
        "description": "Cloud Run OIDC token presented by Google Cloud Tasks in production.",
        "scheme": "bearer",
        "type": "http"
      },
      "CollectorBearer": {
        "description": "Collector credential returned by /collector/v1/register or /collector/v1/credentials/rotate.",
        "scheme": "bearer",
        "type": "http"
      },
      "DevelopmentIdentityHeader": {
        "description": "Local development/test identity selector. Disabled in production.",
        "in": "header",
        "name": "x-archagent-user",
        "type": "apiKey"
      },
      "WorkOSSessionCookie": {
        "description": "Sealed WorkOS AuthKit session cookie used in hosted auth mode.",
        "in": "cookie",
        "name": "__Host-archagent-session",
        "type": "apiKey"
      },
      "WorkOSWebhookSignature": {
        "description": "WorkOS webhook signature header validated with the configured webhook secret.",
        "in": "header",
        "name": "workos-signature",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "\nArchAgent API contract for browser frontend, collector, staff, webhook, and worker integrations.\n\nUse `/openapi.json` as the machine-readable source of truth. Swagger UI is available at `/docs`\nand ReDoc is available at `/redoc` when the FastAPI app is running.\n",
    "title": "ArchAgent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/analysis-runs": {
      "get": {
        "description": "List analysis runs for the active organization, optionally filtered by cluster_id.",
        "operationId": "list_analysis_runs_v1_analysis_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "cluster_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AnalysisRunResponse"
                  },
                  "title": "Response List Analysis Runs V1 Analysis Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "List Analysis Runs",
        "tags": [
          "Product"
        ]
      },
      "post": {
        "description": "Queue or run an analysis for the active organization. Requires owner or admin role.",
        "operationId": "create_analysis_run_v1_analysis_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "CSRFHeader": [],
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "Create Analysis Run",
        "tags": [
          "Product"
        ]
      }
    },
    "/v1/analysis-runs/{run_id}": {
      "get": {
        "description": "Return one analysis run by id for the active organization.",
        "operationId": "get_analysis_run_v1_analysis_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "Get Analysis Run",
        "tags": [
          "Product"
        ]
      }
    },
    "/v1/audit-log": {
      "get": {
        "description": "Return the latest 100 organization audit events for the active organization.",
        "operationId": "get_audit_log_v1_audit_log_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEventResponse"
                  },
                  "title": "Response Get Audit Log V1 Audit Log Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "List Audit Log",
        "tags": [
          "Product"
        ]
      }
    },
    "/v1/clusters": {
      "get": {
        "description": "List Kubernetes clusters owned by the active organization.",
        "operationId": "get_clusters_v1_clusters_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClusterResponse"
                  },
                  "title": "Response Get Clusters V1 Clusters Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "List Clusters",
        "tags": [
          "Product"
        ]
      },
      "post": {
        "description": "Create a Helm-connected cluster record for the active organization. Requires owner or admin role.",
        "operationId": "add_cluster_v1_clusters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClusterCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "CSRFHeader": [],
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "Create Cluster",
        "tags": [
          "Product"
        ]
      }
    },
    "/v1/clusters/{cluster_id}/namespaces": {
      "get": {
        "description": "List namespace monitoring choices for a cluster.",
        "operationId": "get_cluster_namespaces_v1_clusters__cluster_id__namespaces_get",
        "parameters": [
          {
            "in": "path",
            "name": "cluster_id",
            "required": true,
            "schema": {
              "title": "Cluster Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NamespaceResponse"
                  },
                  "title": "Response Get Cluster Namespaces V1 Clusters  Cluster Id  Namespaces Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "List Cluster Namespaces",
        "tags": [
          "Product"
        ]
      },
      "put": {
        "description": "Replace namespace monitoring choices for a cluster. Requires owner or admin role.",
        "operationId": "set_cluster_namespaces_v1_clusters__cluster_id__namespaces_put",
        "parameters": [
          {
            "in": "path",
            "name": "cluster_id",
            "required": true,
            "schema": {
              "title": "Cluster Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NamespaceUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NamespaceResponse"
                  },
                  "title": "Response Set Cluster Namespaces V1 Clusters  Cluster Id  Namespaces Put",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "CSRFHeader": [],
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "Replace Cluster Namespaces",
        "tags": [
          "Product"
        ]
      }
    },
    "/v1/clusters/{cluster_id}/registration-token": {
      "post": {
        "description": "Rotate and return a one-time collector registration token for Helm installation. Requires owner or admin role.",
        "operationId": "create_collector_registration_token_v1_clusters__cluster_id__registration_token_post",
        "parameters": [
          {
            "in": "path",
            "name": "cluster_id",
            "required": true,
            "schema": {
              "title": "Cluster Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectorRegistrationTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "CSRFHeader": [],
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "Create Collector Registration Token",
        "tags": [
          "Product"
        ]
      }
    },
    "/v1/knowledge/documents": {
      "get": {
        "description": "List organization-scoped knowledge documents available to the active organization.",
        "operationId": "get_documents_v1_knowledge_documents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  },
                  "title": "Response Get Documents V1 Knowledge Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "List Knowledge Documents",
        "tags": [
          "Product"
        ]
      }
    },
...
    "/v1/session": {
      "get": {
        "description": "Return the authenticated user, active organization, role, permissions, and selectable organizations.",
        "operationId": "get_session_v1_session_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "Get Customer Session",
        "tags": [
          "Product"
        ]
      }
    },
...
  },
...
}